Refactoring Code#1
Create LargeCard and SmallCard components

Made a stateless component for both large card and small card.
This seperates the toggling functionality and the rendering
conerns. Small and large card are also not reusable.

This change was necessary to maintain readability of code. This will
also make both cards reusable in other components, if necessary.

Create Nav container and change NavBar

Nav container created to seperate handleScrollToElement function
from NavBar rendering. Now NavBar.js is a stateless component and
no longer has handleScrollToElement function. handleScrollToElement
is now located inside Nav.js.

This was done to maintain code readability and reusability.

Create SocialNetworksButtons

Added SocialNetworksButtons component and removed social networks
from AboutMe.js and Footer.js. AboutMe.js and Footer.js now uses
SocialNetworksButtons component.

SocialNetworksButtons was created due to repeating code in AboutMe
and Footer. This allows for clean and reusable code.

Create Particles component

Extracted Particles component from App.js into a seperate
component file.

Moving Particles into its own component improves readability of
App.js and makes it easier to change params of Particles.
