+++ src/sass/components/Portfolio/Portfolio.scss
@@ -7,7 +7,6 @@
 	justify-content: flex-start;
 	align-items: center;
 	width: 100%;
-	// background-color: $color-scheme-three;
 	margin: 100px auto;
 
 	h2{
